pyproject.toml
[project]
name = "edyouapp-api"
version = "1.0.0"
description = "EDYOU Api project"
authors = [
    { name = "Hassnat", email = "hasnat@o2geeks.com" }
]
dependencies = [
    "sentry-sdk[fastapi]>=2.6.0",
    "pydantic-settings>=2.3.3",
    "motor>=3.4.0",
    "pyjwt>=2.8.0",
    "bugsnag>=4.7.1",
    "loguru>=0.7.2",
    "livekit-api>=0.5.1",
    "httpx>=0.27.0",
    "redis>=5.0.6",
    "email-validator>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "fastapi-slim>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "sqlalchemy>=2.0.32",
    "requests>=2.32.3",
    "boto3>=1.35.2",
    "python-multipart>=0.0.9",
    "pyapns-client3>=3.0.6",
    "apscheduler>=3.10.4",
    "slixmpp>=1.10.0",
    "pika>=1.3.2",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "factory-boy>=3.3.0",
]

[tool.rye.scripts]
devserver = { cmd = "uvicorn edyouapp_api.app:app --reload --port 8080 --log-level debug", env-file = "env.local"}

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["edyouapp_api"]

[tool.ruff]
line-length = 88
indent-width = 4
# Assume Python 3.12
target-version = "py312"
[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B904",
    "E722"
]
preview = true

[tool.ruff.lint.isort]
case-sensitive = true
combine-as-imports = true
known-third-party = ["fastapi", "pydantic", "starlette"]
[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
